Adds CI workflow for fork pull requests
Sets up a comprehensive CI workflow specifically for pull requests originating from forks.

This workflow includes jobs for:
- Formatting and linting the code.
- Running cargo check with different feature flag combinations.
- Checking the verifier in a no_std environment.
- Checking the test-artifacts package.
- Checking the examples.

The workflow is triggered on pull requests to the `dev` and `thep2p-dev` branches, when changes are made to crates, examples, the Cargo.toml file, or workflow files.

Also installs Go version 1.22 for all the jobs.